Proyectos/Tool de Identificación de Requisitos de Software/backend_tool/requirements/base.txt
django==2.2.3
django-cors-headers==2.3.0
django-environ==0.4.5
djangorestframework==3.8.2
psycopg2-binary==2.8.5
drf-yasg==1.10.0
Pillow==7.2.0
jsonfield==2.0.2